clawstack_v2/docker/workstudy_app/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gradio>=4.0.0
mediapipe>=0.10.8
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.1.0
pyyaml>=6.0
reportlab>=4.0.0
openpyxl>=3.1.0
python-multipart>=0.0.6
